+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* basic CLI
+* specifying columns
+* adding foreign keys
+
 ### Fixed
 * adding more than just 1 entry